sandbox-classic: Split out `LedgerBackedWriteService`.
